Created separate main function for tests
