+++ test/000-test-main.cc
@@ -1,0 +1,3 @@
+#define CATCH_CONFIG_MAIN
+#include "catch.hpp"
+
